+++ model/CHANGELOG.md
@@ -4,11 +4,15 @@ This file documents changes to the `workbench-model` library, including notes on
 
 ## 0.14
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-model" % "0.14-3c0b510"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-model" % "0.14-TRAVIS-REPLACE-ME"`
 
 ### Changed
 
 - added IdentityConcentratorId to WorkbenchUser 
+
+### Added
+
+- `IP` data type
 
 ## 0.13